Copy the full results directory after each CB run
The "latest" directory no longer exists, so we can't rely on that.  We
could try to find the name of the specific directory for the latest run,
but why bother?  The results directory is recreated for each run so we
know it only has the files we care about.

The results.json file will be copied along with the logs, so there is no
longer a need to email that file separately.